# Yum/src/main/resources/static/api/swagger-ui/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.2.7",
        "title": "Yum Food Orders",
        "license": {
            "name": "GPLv3",
            "url": "http://www.gnu.org/licenses/"
        },
        "description": "**Yum application, order food daily from the best chef in town**\n\nThis API is used by the angular.io client, and is not meant to be used otherwise.\n\nFind source code of this API [here](http://gitlab/)\n\nCopyright (C) 2017 JR Technologies.\n\n  \n------------------------------------\n\n    Last edit: 18/05/2017 13:30\n\n-------------------------------------\n"
    },
    "basePath": "/api",
    "schemes": [
        "http",
        "https"
    ],
    "produces": [
        "application/json"
    ],
    "securityDefinitions": {
        "Bearer": {
            "description": "For accessing the API a valid JWT token must be passed in all the queries in the 'Authorization' header.\nA valid JWT token is generated by the API and returned as answer of a call to the path /api/auth/login giving a valid user & password.\nThe following syntax must be used in the 'Authorization' header :\nBearer xxxxxx.yyyyyyy.zzzzzz\n",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "paths": {
...
        "/foods": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "chef"
                ],
                "summary": "Gets all foods, optionally return stats per food",
                "description": "Return a list of all foods",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "stats",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "in": "query",
                        "name": "page",
                        "description": "Request pagination page",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "in": "query",
                        "name": "size",
                        "description": "Request pagination size / num of foods",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "foodType",
                        "in": "query",
                        "description": "Request foodType filter",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "archived",
                        "in": "query",
                        "description": "Request archived filter",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "orderBy",
                        "in": "query",
                        "description": "Request orderBy filter",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "orderDirection",
                        "in": "query",
                        "description": "Request orderBy filter",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of foods",
                        "schema": {
                            "$ref": "#/definitions/FoodsPage"
                        }
                    },
                    "500": {
                        "$ref": "#/responses/Error"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "chef"
                ],
                "summary": "A new food",
                "description": "Create a food with basic data",
                "parameters": [
                    {
                        "name": "foodDetails",
                        "in": "body",
                        "description": "The food to save",
                        "schema": {
                            "properties": {
                                "foodName": {
                                    "type": "string"
                                },
                                "foodType": {
                                    "type": "string"
                                },
                                "description": {
                                    "type": "string"
                                },
                                "price": {
                                    "type": "number",
                                    "format": "double"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Food succesfully created"
                    },
                    "400": {
                        "$ref": "#/responses/Error"
                    },
                    "412": {
                        "description": "Food name already exists",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "$ref": "#/responses/Error"
                    }
                }
            }
        },
        "/foods/findByName/{name}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "chef"
                ],
                "summary": "Get food by name",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "name",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Food found",
                        "schema": {
                            "$ref": "#/definitions/Food"
                        }
                    },
                    "404": {
                        "description": "Food not found"
                    },
                    "500": {
                        "$ref": "#/responses/Error"
                    }
                }
            }
        },
        "/foods/{foodId}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "chef"
                ],
                "summary": "Get food by id, optionally check if editable",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "foodId",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "in": "query",
                        "name": "editable",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Food get ok",
                        "schema": {
                            "$ref": "#/definitions/FoodEditable"
                        }
                    },
                    "404": {
                        "description": "Food not found"
                    },
                    "500": {
                        "$ref": "#/responses/Error"
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "chef"
                ],
                "summary": "Updates an existing food",
                "description": "Chef can update food name, type, description or price",
                "parameters": [
                    {
                        "in": "path",
                        "name": "foodId",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "in": "query",
                        "name": "clone",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "food",
                        "in": "body",
                        "description": "The food to be updated",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/EditedFood"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "$ref": "#/responses/MessageResponse"
                    },
                    "400": {
                        "$ref": "#/responses/Error"
                    },
                    "404": {
                        "description": "Food not found"
                    },
                    "409": {
                        "description": "Concurrent modification error",
                        "schema": {
                            "$ref": "#/definitions/EditedFood"
                        }
                    },
                    "412": {
                        "description": "Food name already exists",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "$ref": "#/responses/Error"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "chef"
                ],
                "summary": "Deletes an existing food",
                "description": "Chef can update food name, type, description or price",
                "parameters": [
                    {
                        "in": "path",
                        "name": "foodId",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "in": "query",
                        "name": "archive",
                        "required": false,
                        "type": "boolean"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Food successfully archived"
                    },
                    "204": {
                        "description": "Food successfully deleted"
                    },
                    "404": {
                        "description": "Food not found"
                    },
                    "412": {
                        "description": "Food couldn't be deleted"
                    },
                    "500": {
                        "$ref": "#/responses/Error"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Error": {
            "type": "object",
            "description": "An error with some info DTO",
            "properties": {
                "message": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                }
            }
        },
        "FoodsPage": {
            "type": "object",
            "properties": {
                "totalElements": {
                    "type": "integer",
                    "format": "int64"
                },
                "totalPages": {
                    "type": "integer",
                    "format": "int64"
                },
                "foods": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/FoodWithStats"
                    }
                }
            }
        },
        "FoodWithStats": {
            "type": "object",
            "description": "Food, statistics  included DTO",
            "properties": {
                "foodItem": {
                    "$ref": "#/definitions/Food"
                },
                "stats": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/Statistic"
                    }
                },
                "editable": {
                    "type": "boolean"
                },
                "lastEdit": {
                    "$ref": "#/definitions/LastEdit"
                }
            }
        },
...
        "FoodEditable": {
            "type": "object",
            "description": "Food,  checked if editable",
            "properties": {
                "foodItem": {
                    "$ref": "#/definitions/Food"
                },
                "editable": {
                    "type": "boolean"
                },
                "lastEdit": {
                    "$ref": "#/definitions/LastEdit"
                }
            }
        },
        "Statistic": {
            "type": "object",
            "description": "Food statistics",
            "properties": {
                "week": {
                    "type": "integer",
                    "format": "int32"
                },
                "quantity": {
                    "type": "integer",
                    "format": "int32"
                }
            }
        },
        "EditedFood": {
            "properties": {
                "foodName": {
                    "type": "string"
                },
                "foodType": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "price": {
                    "type": "number",
                    "format": "double"
                },
                "lastEdit": {
                    "$ref": "#/definitions/LastEdit"
                }
            }
        },
        "Food": {
            "type": "object",
            "description": "Food DTO",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "foodName": {
                    "type": "string"
                },
                "foodType": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "price": {
                    "type": "number",
                    "format": "double"
                },
                "archived": {
                    "type": "boolean"
                }
            }
        },
...
        "LastEdit": {
            "type": "object",
            "description": "Last edit DTO",
            "properties": {
                "timeStamp": {
                    "type": "string",
                    "format": "date-time"
                },
                "version": {
                    "type": "number"
                }
            }
        },
        "Message": {
            "type": "object",
            "description": "A custom message.",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "responses": {
        "Error": {
            "description": "An unexpected error occured.",
            "schema": {
                "$ref": "#/definitions/Error"
            }
        },
        "MessageResponse": {
            "description": "A custom message.",
            "schema": {
                "$ref": "#/definitions/Message"
            }
        }
    }
}